+++ battery_notifier/configs.py
@@ -8,8 +8,8 @@ NOTIFICATION_THRESHOLD = 20
 
 def resource_path(relative_path: str) -> Path:
     """Get absolute path to resource, works for dev and for PyInstaller"""
-    current_path = Path(__file__).parent
-    base_path = getattr(sys, "_MEIPASS", current_path.absolute())
+    current_path = Path(__file__).parent.absolute()
+    base_path = Path(getattr(sys, "_MEIPASS", current_path))
     return base_path / relative_path